Updated names + moved alphabet object
